pyproject.toml
[tool.poetry]
name = "tweets"
version = "0.1.0"
description = ""
authors = ["Jinho Jeong <zinojung@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.1"
djangorestframework = "^3.15.2"
pyjwt = "^2.9.0"
django-environ = "^0.11.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
